use size_t to hold string length
using int is not fine on 64bit systems

+++ src/main.c
@@ -4169,7 +4169,7 @@ static char *get_url_file_name(const char *url)
 }
 
 static char*
-parse_filename(char *ptr, int len)
+parse_filename(char *ptr, size_t len)
 {
   char* copy;
   char* p;
